+++ contaminer/models/tests.py
@@ -1735,7 +1735,9 @@ class JobTestCase(TestCase):
                         'uniprot_id': "P0ACJ8",
                         'status': 'Running',
                         'percent': 40,
-                        'q_factor': 0.53
+                        'q_factor': 0.53,
+                        'pack_number': 1,
+                        'space_group': "P-1-2-1",
                     },
                 ]
             }
@@ -1853,6 +1855,8 @@ class JobTestCase(TestCase):
                         'status': 'Complete',
                         'percent': 50,
                         'q_factor': 0.60,
+                        'pack_number': 1,
+                        'space_group': "P-1-2-1",
                     },
                 ]
             }
@@ -1913,6 +1917,8 @@ class JobTestCase(TestCase):
                         'status': 'Complete',
                         'percent': 40,
                         'q_factor': 0.70,
+                        'pack_number': 1,
+                        'space_group': "P-1-2-1",
                     },
                 ]
             }
@@ -2699,9 +2705,7 @@ class JobTestCase(TestCase):
 
 
 class TaskTestCase(TestCase):
-    """
-        Test the Task model
-    """
+    """Test the Task model"""
     def setUp(self):
         Job.objects.create(
                 name = "test",
@@ -2805,6 +2809,110 @@ class TaskTestCase(TestCase):
         self.assertEqual(
                 Task.from_name(self.job, "P0ACJ8_5_P-2-2-2"),
                 task)
+
+    def test_comp_gives_expected_result_normal_cases(self):
+        task1 = Task.objects.create(
+                job = self.job,
+                pack = self.pack,
+                space_group = "P-2-2-2",
+                percent = 51,
+                q_factor = 0.9,
+                )
+        task2 = Task.objects.create(
+                job = self.job,
+                pack = self.pack,
+                space_group = "P-21-2-2",
+                percent = 52,
+                q_factor = 0.9,
+                )
+        task3 = Task.objects.create(
+                job = self.job,
+                pack = self.pack,
+                space_group = "P-2-21-2",
+                percent = 51,
+                q_factor = 0.9,
+                )
+        task4 = Task.objects.create(
+                job = self.job,
+                pack = self.pack,
+                space_group = "P-2-2-21",
+                percent = 51,
+                q_factor = 0.8,
+                )
+        self.assertEqual(task1.__cmp__(task2), -1)
+        self.assertEqual(task2.__cmp__(task1),  1)
+        self.assertEqual(task1.__cmp__(task3),  0)
+        self.assertEqual(task1.__cmp__(task4),  1)
+
+    def test_comp_gives_expected_result_edge_cases(self):
+        task1 = Task.objects.create(
+                job = self.job,
+                pack = self.pack,
+                space_group = "P-2-2-2",
+                percent = None,
+                q_factor = None,
+                )
+        task2 = Task.objects.create(
+                job = self.job,
+                pack = self.pack,
+                space_group = "P-21-2-2",
+                percent = 1,
+                q_factor = 0,
+                )
+        task3 = Task.objects.create(
+                job = self.job,
+                pack = self.pack,
+                space_group = "P-2-21-2",
+                percent = 1,
+                q_factor = None,
+                )
+        self.assertEqual(task1.__cmp__(task2), -1)
+        self.assertEqual(task2.__cmp__(task1),  1)
+        self.assertEqual(task1.__cmp__(task3), -1)
+        self.assertEqual(task2.__cmp__(task3),  1)
+
+    def test_comp_extends_to_operators(self):
+        task1 = Task.objects.create(
+                job = self.job,
+                pack = self.pack,
+                space_group = "P-2-2-2",
+                percent = 51,
+                q_factor = 0.9,
+                )
+        task2 = Task.objects.create(
+                job = self.job,
+                pack = self.pack,
+                space_group = "P-21-2-2",
+                percent = 52,
+                q_factor = 0.9,
+                )
+        task3 = Task.objects.create(
+                job = self.job,
+                pack = self.pack,
+                space_group = "P-2-21-2",
+                percent = 52,
+                q_factor = 0.9,
+                )
+
+        self.assertTrue(task1 < task2)
+        self.assertFalse(task1 > task2)
+
+    def test_comp_does_not_override_django__eq__(self):
+        task1 = Task.objects.create(
+                job = self.job,
+                pack = self.pack,
+                space_group = "P-21-2-2",
+                percent = 52,
+                q_factor = 0.9,
+                )
+        task2 = Task.objects.create(
+                job = self.job,
+                pack = self.pack,
+                space_group = "P-2-21-2",
+                percent = 52,
+                q_factor = 0.9,
+                )
+        self.assertFalse(task1 == task2)
 
     def test_get_status_gives_good_result(self):
         task = Task.objects.create(